+++ src/dayjs/min-dayjs/min-dayjs.predicate.ts
@@ -1,5 +1,5 @@
-import type { ConfigType, Dayjs } from 'dayjs'
 import dayjs from 'dayjs'
+import type { ConfigType, Dayjs, OpUnitType } from 'dayjs'
 
 import { isDayjs } from '../../dayjs/is-dayjs'
 
@@ -7,14 +7,23 @@ import { isDayjs } from '../../dayjs/is-dayjs'
  * @category Predicates
  * @param value The value to validate.
  * @param minimum The minimum allowed value.
+ * @param options Additional options (see {@link MinDays}).
  */
-export function minDayjs(value: unknown, minimum: ConfigType): value is Dayjs {
+export function minDayjs(
+    value: unknown,
+    minimum: ConfigType,
+    options?: { allow_invalid?: boolean; inclusive?: boolean; granularity?: OpUnitType }
+): value is Dayjs {
     const min = dayjs(minimum)
+    const inclusive = options?.inclusive ?? false
+    const granularity = options?.granularity ?? 'milliseconds'
 
     if (!min.isValid()) {
         throw new TypeError(`Parameter "minimum" must be a valid date`)
     }
 
     // Let's not rely on the isSameOrBefore-plugin which might or might not be registered.
-    return isDayjs(value, { is_valid: true }) && !(min.isAfter(value) || min.isSame(value))
+    return (
+        isDayjs(value, options) && (min.isBefore(value, granularity) || (inclusive && min.isSame(value, granularity)))
+    )
 }