+++ src/minimal.ts
@@ -1,0 +1,6 @@
+export * from './array'
+export * from './bigint'
+export * from './date'
+export * from './map'
+export * from './set'
+export * from './string'